+++ MovieApp/MovieApp/SectionCellView.swift
@@ -1,0 +1,8 @@
+//
+//  SectionCellView.swift
+//  MovieApp
+//
+//  Created by Mateo Piskac on 10.05.2024..
+//
+
+import Foundation
